+++ changelog.md
@@ -1,3 +1,8 @@
 # Changelog
+
+## <cg>v1.0.0 [Beta 2]</c>
+* Added loading spinner during news loading.
+* New updated button logo for cleaner design.
+
 ## <cg>v1.0.0 [Beta 1]</c>
 * Initial beta release for the public.